Add `Bar` to HiltComposableScreen
This commit introduces a new class `Bar` and integrates it into the `HiltComposableScreen`.

The `Bar` class, along with its associated Dagger Hilt components (`BarEntryPoint`, `BarModule`), is created and provided to the `HiltComposableScreen`.

The `HiltComposableScreen` now displays the `name` property of both `Foo` and `Bar` objects.

A `rememberBar` composable function is added to facilitate remembering the `Bar` instance.